requirements.txt
asn1crypto==0.24.0
Babel==2.5.1
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloud-init==18.2
configobj==5.0.6
cryptography==2.3
decorator==4.2.1
ethtool==0.14
file-magic==0.3.0
Flask==1.1.1
Flask-MySQL==1.4.0
gpg==1.10.0
idna==2.5
iniparse==0.4
insights-client==3.0.5
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpatch==1.21
jsonpointer==1.10
jsonschema==2.6.0
MarkupSafe==0.23
netifaces==0.10.6
oauthlib==2.1.0
pciutils==2.3.6
perf==0.1
ply==3.9
prettytable==0.7.2
pycparser==2.14
pygobject==3.28.3
pyinotify==0.9.6
PyJWT==1.6.1
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pyserial==3.1.1
PySocks==1.6.8
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-linux-procfs==0.6
pytz==2017.2
pyudev==0.21.0
PyYAML==3.12
requests==2.20.0
rhnlib==2.8.6
rpm==4.14.2
schedutils==0.6
sepolicy==1.1
setools==4.2.0
six==1.11.0
subscription-manager==1.23.8
syspurpose==1.23.8
urllib3==1.23
virtualenv==16.7.2
Werkzeug==0.15.5
